stacks/requirements.txt
ffmpeg-python==0.2.0
ultralytics==8.1.5
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
imutils
scikit-image